@@ -7,18 +7,17 @@
     "@emotion/react": "^11.5.0",
     "@emotion/styled": "^11.3.0",
     "@mui/material": "^5.0.4",
-    "react": "^17.0.2",
-    "react-dom": "^17.0.2",
-    "react-scripts": "^5.0.1",
+    "react": "^18.3.1",
+    "react-dom": "^18.3.1",
     "typescript": "^4.3.5"
   },
   "scripts": {
-    "start": "react-scripts start",
+    "dev": "vite",
+    "build": "vite build",
+    "preview": "vite preview",
     "predeploy": "npm run build",
-    "deploy": "gh-pages -d build",
-    "build": "react-scripts build",
-    "test": "react-scripts test",
-    "eject": "react-scripts eject",
+    "deploy": "gh-pages -d dist",
+    "test": "jest",
     "lint": "eslint ./src/**/*.{ts,tsx}",
     "lint:fix": "yarn lint --fix"
   },
@@ -43,14 +42,14 @@
     "@iconify-icons/akar-icons": "^1.1.17",
     "@iconify-icons/logos": "^1.1.10",
     "@iconify/react": "^1.1.4",
-    "@types/react": "^17.0.15",
-    "@types/react-dom": "^17.0.9",
-    "@types/styled-components": "^5.1.13",
+    "@types/react": "^18.3.27",
+    "@types/react-dom": "^18.3.7",
+    "@vitejs/plugin-react": "^5.1.1",
     "eslint": "^7.32.0",
     "eslint-config-prettier": "^8.3.0",
-    "eslint-config-react": "^1.1.7",
     "eslint-plugin-prettier": "^4.0.0",
     "gh-pages": "^6.3.0",
-    "prettier": "^2.4.0"
+    "prettier": "^2.4.0",
+    "vite": "^7.2.2"
   }
 }
